Adds the ability to rename and relink pages on title change
A lot had to happen under the hood for this, including refactoring some
of the ways we deal with parsing the wiki location moving from ingesting
and returning a PathBuf to just dealing with String and &str. The
current approach is not optimized, but we can improve upon that in
subsequent commits.