Splits up business logic out from handlers to controllers
It was difficult to read through the large handler functions that mixed
different types of logic, from applying filters to the request (like
`with_auth`), to rendering out HTML. By splitting this out into
"controllers", we can keep the hanlders themselves high-level, acting as
routers which apply certain filters to the request, and finally execute
business logic on filtered request.

Also in this commit (scope creep, I know) is the controller for handling
the upload of binary files via a form. This allows us to attach
documents like PDFs or images to our wiki, which is a much more rich
multi-media experience. These files will not be displayed inside the
rendered wiki (yet), but instead can be viewed in the browser via
linking with a syntax extension (not in the commit, but in one to
come).