Cleans up handlers from a huge list of one-off functions.
This organizes them into structs by domain. All of the handlers related
to editing, deleting, viewing, creating a wiki page are in one struct,
all of the handlers for serving static files in anther, etc.

One thing that still remains, is to organize the lonely function in
"services.rs", as well as moving the templates into the hanlders page.