Makes ParserMachine struct's send method return new state and Event enum
This allows us to explore moving some of the business logic from the
`to_html` function into the state machine, making it easier to read and
reason about which state transistions emit which Pulldown CMark events.
This also help make our parser more deterministic and narrow down bugs
when things go wrong.

Since we now can print out which state the machine currently is in and
which state it should transistion to, we can more easily understand
where parsing fails or if there are any performance bottleneck
 Please enter the commit message for your changes. Lines starting