Changes architecture for updating links and references when notes are updated.
This new architecture is based on MPSC message passing. This allows all
the logic for updating the note cache and the global maps of backlinks
and tags to be in one place.

This also has the advantage that all of our web handlers now only need
read-only access to these in memory stores, giving us more flexibility
for the future since we won't need to do much to appease the borrow
checker.

Another advantage of this change is that we can hand out clones of the
MPSC tx handle so that our ability to update the global stores is not
just limited to the web server process. We can now communicate changes
via the git sync process, and expand this out if needed.