Improve search functionality, consolidate search pages.
This simplifies the two different search functions and search results
pages by doing a context search on every search and unifying the results
in one page. This also moves the search functionality from a thread-pool
to an async task iterator. This change eliminates some of the
complications that come from sharing memory and uses functional iterator
methods to filter_map results.