Parser: Rewrite parser to parse new wikitext syntax.
Simplifying the note taking syntax forces focus back on the content of
the notes rather than the presentation. It also forces Tendril Wiki to
provide functionality for composing different types of data together in
a presentation layer. If we want text notes to be a fundamental building
block, they must be as simple as possible.