Autocomplete: Change method of switching from editor to viewer.
This moves away from element.blur() to explicitly setting the view or
editor as the active element. The reason we need to do this is because
focusing or clicking the autocomplete menu will cause the textarea to
blur, losing your editing context and forcing you to click the
text-block again to continue editing after selecting an item from the
autocomplete menu.