Autocomplete: Integrate scorer into autocomplete flow
When typing, the user is presented a list of potential match candidates
based on the Levenshtein distance between the current word being typed
and all available note titles. The menu behaves similarly to the
autocomplete menu in text editors where 'Tab' moves down the list and
'Enter' selects an item.