Refactor: Split work being done in www into task runners.
The work that was currently being done in the www crate isn't
necessarily related to the web server. Tasks such as updating the cache
file, writing new text to a wiki entry, and deleting a note can receive
their input from an HTTP request, but it's not imperative that they do
so. This change splits up the tasks from the server making it easier to
re-use this logic when/if we move to a desktop GUI app over an app
served over the web.