Search: Improve search results speed
This change involves several different improvements:

1: Instead of constructing a docs_index and search_index and loading
them both into memory when executing a search, this simplifies things to
a single inverse-index.
2. Context for each search result required reading the whole file,
writing it to one of the indices, and then loading _all_ text content
for _all_ notes into memory for a single set of search results! Now
search context is shown on-demand in the UI with a cache-control header
making it easy to toggle between results without having to make the same
network request multiple times.
3. Instead of using a JSON string that gets deserialized into a giant
HashMap, I switched to bincode to make the search index more compact,
thus speeding up the time it takes to load it from disk into memory.
4. Instead of a monolithic search index that needs to be loaded into
memory on each query, I split each term into its own bincode encoded
file containing an array of tuples which hold the documentId and
the frequency of the search term in the documents. These files are
compact and are quickly loaded into memory for scoring.